requirements.txt
customtkinter
pyaudio
requests
cryptography
web3
google-generativeai
google-genai
reportlab
